More opcodes

+++ src/ppu.rs
@@ -324,8 +324,6 @@ impl Ppu {
         let sprite_y = mmu.read_byte(oam_start);
         let sprite_x = mmu.read_byte(oam_start + 1);
 
-        // println!("x: {sprite_x}, y: {sprite_y}");
-
         if sprite_y == 0 || sprite_y >= 160 {
             return;
         }
@@ -384,8 +382,8 @@ impl Ppu {
                     continue;
                 }
 
-                let screen_x = start_x.wrapping_add(x as i16);
-                let screen_y = start_y.wrapping_add(y as i16);
+                let screen_x = start_x.wrapping_add(x as i16) as u8;
+                let screen_y = start_y.wrapping_add(y as i16) as u8;
 
                 if !Self::is_on_screen(screen_x as u8, screen_y as u8) {
                     continue;
